removed the questions controller and moved some of those methods to the interactions controller, moved some of the code in the Api controller into breach checker service so that my controlleris cleaner and the service habdles the key calling and handling errors
