+++ app/views/api/check_breaches.html.erb
@@ -13,7 +13,7 @@
               <!-- Description -->
               <h3 class="text-light heading3 card-title  mb-0">Ask Custos if you need help on what actions to take next!</h3>
               <!-- Buttons -->
-              <%= link_to "Ask Custos", new_question_path, class: "btn btn-primary heading3 border-light  text-light  mt-3" %>
+              <%= link_to "Ask Custos", new_interaction_path, class: "btn btn-primary heading3 border-light  text-light  mt-3" %>
             </div>
           
         </div>
