+++ app/views/interactions/_form.html.erb
@@ -1,4 +1,4 @@
-<!-- app/views/questions/_form.html.erb -->
+<!-- app/views/interactions/_form.html.erb -->
 <%= form_with url: ask_question_path, local: true, class: "d-flex flex-column align-items-start" do |form| %>
   <div class="mb-3 w-100">
     <%= form.label :question, "Ask Custos a Question:", class: "form-label text-dark" %>
