+++ app/views/interactions/new.html.erb
@@ -1,6 +1,6 @@
 
-  <div class="container" 
-    <div class="container m-3" 
+  <div class="container" >
+    <div class="container m-3" >
   <!-- Image Card for Custos -->
   <div class="card bg-white mb-4" style="max-width: 25rem; margin: 0 auto; border: none; border-radius: 15px;">
     <img src="<%= asset_path('askbot.gif') %>" alt="Custos the Cybersecurity Expert Bot" class="card-img-top">
@@ -12,7 +12,7 @@
     <h3 class="card-title text-center" style="margin-bottom: 0; font-size: 1.5rem;">Have a CyberSecurity Question? </h3>
   </div>
   <div class="card-body m-2" style="background-color: white; padding-top: 0; border-bottom-left-radius: 1.5rem; border-bottom-right-radius: 1.5rem;">
-    <%= render 'questions/form' %>
+    <%= render 'interactions/form'  %>
   </div>
 </div>
