+++ app/views/interactions/show.html.erb
@@ -1,17 +1,23 @@
-<!-- app/views/interactions/show.html.erb -->
+<!-- app/views/questions/show.html.erb -->
 <div class="container mt-4">
-  <div class="card border-dark mb-4" style="border-radius: 1.5rem; margin: 0 auto; max-width: 60vw;">
-    <div class="card-header bg-dark text-light" style="padding: 1rem; border-top-left-radius: 1.5rem; border-top-right-radius: 1.5rem;">
-      <h3 class="card-title text-center " style="margin-bottom: 0; font-size: 2rem;">Question Asked on <%= @interaction.created_at.strftime("%b %d, %Y") %></h3>
+  <!-- Enhanced Image Card for Custos -->
+  <div class="card bg-white mb-4" style="max-width: 25rem; margin: 0 auto; border: none; border-radius: 15px;">
+    <img src="<%= asset_path('custos.gif') %>" alt="Custos the Cybersecurity Expert Bot" class="card-img-top">
+    <div class="card-body bg-dark text-info" style="padding: 1rem; border-radius: 15px;">
+      <h5 class="card-title mb-0 text-center">Custos the CyberSecurity Expert</h5>
     </div>
-    <div class="card-body" style="background-color: white; padding-top: 0; border-bottom-left-radius: 1.5rem; border-bottom-right-radius: 1.5rem;">
-      <div class="mb-4">
-        <h4 class="card-title" style="margin-top: 1rem;">Question:</h4>
-        <p class="card-text" style="font-size: 1.4rem; line-height: 1.6;"><%= @interaction.question %></p>
-        <hr>
-        <h4 class="card-title">Answer:</h4>
-        <p class="card-text" style="font-size: 1.4rem; line-height: 1.6; white-space: pre-wrap;"><%= @interaction.answer %></p>
-      </div>
+  </div>
+
+  <!-- Customized Question and Answer Card -->
+  <div class="card border-dark mb-4" style="border-radius: 15px; margin: 0 auto; max-width: 60vw;">
+    <div class="card-header bg-dark text-info" style="padding: 1rem; border-top-left-radius: 15px; border-top-right-radius: 15px;">
+      <h3 class="card-title text-center" style="margin-bottom: 0; font-size: 2rem;">Hello <%= @user.name.split.map(&:capitalize).join(' ') %></h3>
+    </div>
+    <div class="card-body" style="background-color: white; padding-top: 0; border-bottom-left-radius: 15px; border-bottom-right-radius: 15px;">
+      <h3 class="card-title" style="margin-top: 1rem;">Question:</h3>
+      <p class="card-text" style="font-size: 1.4rem; line-height: 1.6;"><%= @interaction.question %></p>
+      <h3 class="card-title">Custos's Response:</h3>
+      <p class="card-text" style="font-size: 1.4rem; line-height: 1.6; white-space: pre-wrap;"><%= @interaction.answer %></p>
     </div>
   </div>
 </div>
