+++ app/views/shared/_navbar.html.erb
@@ -26,7 +26,7 @@
         <ul class="navbar-nav justify-content-end flex-grow-1 pe-3">
           <!-- Custos Navbar Item -->
           <li class="nav-item">
-          <%= link_to "Ask Custos", new_question_path, class: "nav-link" %>
+          <%= link_to "Ask Custos", new_interaction_path, class: "nav-link" %>
           </li>
 
           <!-- My Emails Navbar Item (Conditional) -->
