+++ app/views/users/show.html.erb
@@ -154,7 +154,7 @@
       <div id="collapseFour" class="accordion-collapse collapse" aria-labelledby="headingFour" data-bs-parent="#accordionExample">
         <div class="accordion-body bg-dark bg-opacity-10 text-dark">
           <div class= "container m-2">
-         <%= render 'questions/form' %>  
+         <%= render 'interactions/form' %>  
         </div>
         </div>
       </div>
